requirements-dev.txt
autopep8==1.5.5
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
flake8==3.9.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
mccabe==0.6.1
oauth2client==4.1.3
packaging==20.9
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDrive2==1.8.1
pyflakes==2.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
